Modernize analyzer for C# 14 patterns
Rename the package to ExhaustiveMatching.Modern, add finite-domain coverage for modern patterns, update the build and documentation, and provide validated NuGet packaging and publishing workflows.

+++ ExhaustiveMatching.Analyzer.Enums/Semantics/SpecialTypeExtensions.cs
@@ -23,9 +23,9 @@ namespace ExhaustiveMatching.Analyzer.Enums.Semantics
                 case SpecialType.System_UInt32:
                     return TypeCode.UInt32;
                 case SpecialType.System_Int64:
-                    return TypeCode.Int32;
+                    return TypeCode.Int64;
                 case SpecialType.System_UInt64:
-                    return TypeCode.UInt32;
+                    return TypeCode.UInt64;
 
                 // Floating Point Types
                 case SpecialType.System_Single:
